requirements.txt
Django==3.1
psycopg2-binary==2.8
python-dotenv==0.15
djangorestframework==3.12
django-filter==2.4
pyyaml==5.3
uritemplate==3.0
python-memcached==1.59
